Create common methods Message.fromCtx() and Message.fromRabbitMessage();
